command_line_arguments: size_t ac

+++ command_line_arguments.hpp
@@ -2,7 +2,7 @@
 
 namespace putils {
     template<typename T>
-    T parseArguments(int ac, const char * const * av) noexcept;
+    T parseArguments(size_t ac, const char * const * av) noexcept;
 }
 
 #include "command_line_arguments.inl"
